Feat: Scroller応急処置
