--- Assets/Project/Commons/UI/Scripts/View/ButtonList.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 67838628027717749aa2b3e26483dd82
