+++ Assets/Project/Commons/UI/Scripts/View/ButtonListBase.cs
@@ -7,12 +7,9 @@ namespace Project.Commons.UI.Scripts.View
 {
     public abstract class ButtonListBase : MonoBehaviour
     {
-        [SerializeField] protected List<ArchivedButtonBase> buttons;
+        [SerializeField] protected List<MovableButton> buttons;
 
         protected ButtonListType buttonListType;
-        
-        readonly Subject<int> onButtonChanged = new();
-        public IObservable<int> OnButtonChanged => onButtonChanged;
 
         public int ButtonIndex { get; protected set; }
         public bool IsActive { get; protected set; }
@@ -37,7 +34,6 @@ namespace Project.Commons.UI.Scripts.View
 
             if (MoveNextFlag) MoveNext();
             if (MoveBackFlag) MoveBack();
-            if (Input.GetKeyDown(KeyCode.Space)) PressButton();
         }
         
         void MoveBack() => MoveNext(false);
@@ -48,8 +44,6 @@ namespace Project.Commons.UI.Scripts.View
             this.buttonListType = buttonListType;
             
             SetButtonIndex(index);
-            buttons[ButtonIndex].SetActive(true);
-            
             SetActive(isActive);
         }
         
@@ -58,30 +52,10 @@ namespace Project.Commons.UI.Scripts.View
             IsActive = active;
         }
 
-        public void SetActiveButton(int index, bool isActive = false)
-        {
-            SetButtonIndex(index);
-            buttons[ButtonIndex].SetActive(isActive);
-        }
-        
-        public IObservable<Unit> GetButtonEvent(int index)
-        {
-            return buttons[index].OnPressed;
-        }
-        
-        public void PressButton() => PressButton(ButtonIndex);
-        public void PressButton(int index)
-        {
-            if (!IsActive) return;
-            
-            buttons[index].Press();
-        }
-        
         public abstract void MoveNext(bool isUp = true);
         
         protected void SetButtonIndex(int index)
         {
-            onButtonChanged.OnNext(index);
             ButtonIndex = (index + buttons.Count) % buttons.Count;
         }
     }
