+++ Assets/Project/Scenes/StageList/Scripts/View/StageCardListView.cs
@@ -1,7 +1,10 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using DG.Tweening;
+using NUnit.Framework;
 using Project.Commons.UI.Scripts.View;
+using UniRx;
 using UnityEngine;
 
 
@@ -10,15 +13,22 @@ namespace Project.Scenes.StageList.Scripts.View
     public class StageCardListView : MonoBehaviour
     {
         [SerializeField] ScrollableButtonList scrollableButtonList;
+        [SerializeField] List<SimpleButton> simpleButtons;
         [SerializeField] SpriteRenderer charaImage;
         
         public List<StageCardView> stageCardViews;
-        public IObservable<int> OnButtonChanged => scrollableButtonList.OnButtonChanged;
+        
+        readonly Subject<int> onButtonChanged = new();
+        public IObservable<int> OnButtonChanged => onButtonChanged;
 
         public void Init()
         {
             scrollableButtonList.Init(ButtonListType.Vertical, isActive: true);
-            
+
+            for (int i = 0; i < simpleButtons.Count; i++)
+            {
+                simpleButtons[i].OnFocusedEvent.Subscribe(PublishOnButtonChanged);
+            }
         }
 
         public void SetCharaImage(Sprite charaSprite)
@@ -27,6 +37,12 @@ namespace Project.Scenes.StageList.Scripts.View
             charaImage.DOFade(0f, 0f);
             charaImage.sprite = charaSprite;
             charaImage.DOFade(1f, 0.25f).SetDelay(0.1f);
+        }
+
+        void PublishOnButtonChanged(Unit _)
+        {
+            var index = simpleButtons.FindIndex(b => b.IsFocused);
+            onButtonChanged.OnNext(index);
         }
     }
 }
